### 2018-01-17 #537 @whedon
[PRE REVIEW]: psrqpy: a python interface for querying the ATNF pulsar catalogue
**Submitting author:** @mattpitkin (<a href="http://orcid.org/0000-0003-4548-526X">Matthew Pitkin</a>)
**Repository:** <a href="https://github.com/mattpitkin/psrqpy" target ="_blank">https://github.com/mattpitkin/psrqpy</a>
**Version:** v0.4.3
