### 2018-01-17 7fd6e50
paper.md: fix affiliations, refs openjournals/joss-reviews#537
