### 2018-01-17 #537 @arfon
@mattpitkin - YAML is whitespace sensitive, it looks like the `index: 1` field on your affiliation isn't aligned with the `name` key above.
