### 2018-01-17 7a3f68e
paper.md: fix key allignment, refs openjournals/joss-reviews#537
