+++ paper/paper.md
@@ -9,7 +9,7 @@ authors:
     affiliation: 1
 affiliations:
   - name: Institute for Gravitational Research, SUPA, University of Glasgow, University Avenue, Glasgow, UK, G12 8QQ
-   index: 1
+    index: 1
 date: 16 January 2018
 bibliography: paper.bib
 ---
