### 2018-01-17 #537 @mattpitkin
@arfon thanks, I was aligning with the colon rather than the start of the key name (I think alignment with the colon is how things are in the [example](http://joss.theoj.org/about#author_guidelines), so that might need to be fixed). I've fixed this in [7a3f68e6](https://github.com/mattpitkin/psrqpy/commit/7a3f68e6ff0783d73e59c11a16828c4d626f4456), and will try recompiling.
