### 2018-01-17 #537 @mattpitkin
A few potential reviewers could be @amrutajaodand who has used the code, or (slightly randomly selected from a set of other pulsar astronomers on github!) @sixbynine, @astrophysically (who has done something vaguely similar [here](https://github.com/astrophysically/ATNF-Pulsar-Cat)), or @demorest. From the list of current reviewers @ygrange might be best.
