### 2018-01-17 #537 @arfon
> @arfon thanks, I was aligning with the colon rather than the start of the key name (I think alignment with the colon is how things are in the example, so that might need to be fixed). I've fixed this in 7a3f68e6, and will try recompiling.

My bad. I've fixed the example docs now.
