### 2018-01-18 #537 @arfon
> Certainly wouldn't mind reviewing this one; but it'd be my first; are there reviewer instructions?

Sure thing. You can find them here: http://joss.theoj.org/about#reviewer_guidelines
