### 2018-01-18 548c957
README.md: add JOSS paper status badge
