### 2018-01-18 8135ead
add function to download the entire catalogue  - a function, get_catalogue, has been added that downloads the    entire ATNF catalogue database and returns it as an astropy    table. This currently does not return the parameter errors.  - refs #3
