### 2018-01-24 f4b02b6
utils.py: add return values to doc strings
