### 2018-01-26 #538 @ygrange
Can one of the editors elaborate on the meaning/background of the question: "Version: Does the release version given match the GitHub release (v0.4.3)?"
Seems like the author has released 0.4.4 by now byt 0.4.3 is still an existing release, even though the paper is part of the master repository and will therefore be part of any future release (which I don't think is a bad thing per se by the way).
