### 2018-01-26 #538 @arfon
@ygrange - we just want to make sure we're reviewing something _close_ to the current version. I think it's fine to go ahead and look at the `v0.4.4` release for your review. Also, yes, the version number often updates as a result of the review feedback (and the fact that a paper is now present in the repository).
