### 2018-01-26 #538 @mattpitkin
@ygrange Sorry, I had meant to say that I'd made a minor update and updated the release number. I added one function that allows you to download the whole ATNF catalogue and return it as an astropy table.
