### 2018-01-26 #538 @ygrange
Thanks for the clarification @mattpitkin !
Another question to the editor. Is it ok from an ethical perspective for me do do a pull request on the repository if I find an issue I can easily fix?
