### 2018-01-30 mattpitkin/psrqpy#12 @ygrange
Last astropy version doesn't build in python2
When building I bumped into an issue I had for another tool before: the current development version of astropy doesn't build with python 2 but is still the default being downloaded if you don't specify anything in the requirements.txt when using python setup.py.

A simple fix, for now at least, is to replace "astropy" by "astropy<3.0" as in that file and it builds like a charm. When using pip, either in the way you use in in travis, or as in the documentation works well. I woud actually advice you to make sure that the installation in travis and the installation instructions in the README are the same.
