### 2018-01-30 #538 @ygrange
Installation: seems to fail when using python2.7 because of astropy dependency. See ticket:
https://github.com/mattpitkin/psrqpy/issues/12
Issue seems small to fix. Will continue with the quick fix I came up with.
