### 2018-01-30 876fee8
Update requirements file to use astropy<3.0:  - versions 3.0 or greater of astropy only support Python 3  - refs #12
