+++ docs/source/index.rst
@@ -34,7 +34,7 @@ The requirements for installing the code are:
  * :mod:`requests`
  * :mod:`bs4`
  * :mod:`numpy`
- * :mod:`astropy`
+ * :mod:`astropy<3.0` (versions 3.0 or greater will not work for Python 2)
  * :mod:`datetime`
 
 The :mod:`ads` module and :mod:`matplotlib` are optional requirement to get the full functionality.
