### 2018-01-30 mattpitkin/psrqpy#12 @mattpitkin
Thanks, I'll make this fix.
