### 2018-01-30 f101d1d
Update requirements for astropy using environment markers
 - refs #12
