### 2018-01-30 0221581
Minor changes to requirements file again
 - refs #12
