+++ requirements.txt
@@ -2,6 +2,6 @@ six
 requests
 beautifulsoup4
 numpy
-astropy<3.0; python_version<='2.7'
-astropy; python_version>'3.4'
+astropy<3.0; python_version <= 2.7
+astropy; python_version > 3.4
 datetime
