### 2018-01-30 8e1c6f8
Update index.rst
 - refs #12
