### 2018-01-30 b8aeb01
Bumped version number fo include updated requirements file in PyPI release - refs #12
