+++ psrqpy/__init__.py
@@ -2,7 +2,7 @@
 
 """ A Python tool for interacting with the ATNF pulsar catalogue """
 
-__version__ = "0.4.4"
+__version__ = "0.4.5"
 
 from .search import QueryATNF
 from .pulsar import Pulsar, Pulsars
