### 2018-01-30 mattpitkin/psrqpy#12 @mattpitkin
This should be fixed now.
