### 2018-01-31 #538 @ygrange
The paper is short and the tool is not very complex. I think this is typically one of those relatively small tools that can turn out to be very useful for people in the community. I have a few comments:

* I couldn't chech the "community guidelines" box yet. Could you read what it states and add such a statement? I guess a single sentence in the README claiming that people wanting to contribute can do a pull request or open a ticket. Also I would add a sentence claiming that all code contributed in this way will adhere to the open source license of the project.
* The documentation does not explicitly state what the exact target audience is and what problem it tries to solve. To me as a reviewer it is rather obvious from the package description what those two are (target audience: astronomers who want to use data from the ATNF pulsar catalogue in scripts; problem to solve: a web interface to the catalog is not exactly the best way to interact with the catalog from scripts; you want a better interface, with a stable API). It may be sensible for non-astronomer readers of the journal to add a sentence of two on this in both the article and the documentation.
* You may want to put an example where the input and output are printed in the README. Alternatively it would be nice to have an integration test that essentially queries a well-known object with one or a few properties that aren't changing anytime son and compare the output to a reference (this could then be added to the tests you do in Travis).
* I can't find an example of the usage of adsref. Could you add one?
* The readthedocs is complete and the API is quite well-described. What I miss here is a description of the classes (Querying, Pulsars, and utility functions; the Setup constants seems obvious to me). Could you add a few sentences describing in short what each the class is meant to do so?
