### 2018-01-31 #538 @mattpitkin
@ygrange - thanks very much for the useful feedback. I'll try and address your points as soon as possible.
