### 2018-02-07 ec12e65
Add statements about contributions and issue reporting to documentation  - refs openjournals/joss-reviews#538
