### 2018-02-07 #538 @mattpitkin
I've [added](https://github.com/mattpitkin/psrqpy/commit/ec12e6561a5920daca3e50ad5d3691c37f2ef24d) a statement about development and support to both the README and documentation.
