### 2018-02-07 983a57b
Add intended user statement  - refs openjournals/joss-reviews#538
