+++ paper/paper.md
@@ -17,10 +17,11 @@ bibliography: paper.bib
 # Summary
 
 This Python module provides an interface for querying the [Australia Telescope
-National Facility (ATNF) pulsar catalogue](http://www.atnf.csiro.au/people/pulsar/psrcat/) [@ATNF]. It allows users to access
+National Facility (ATNF) pulsar catalogue](http://www.atnf.csiro.au/people/pulsar/psrcat/) [@ATNF].
+The intended users are astronomers wanting to extract data from the catalogue through a
+script rather than having to download and parse txet tables output using the standard web interface. It allows users to access
 information, such as pulsar frequencies and sky locations, on all pulsars in
-the catalogue without having to use the current web interface. As such,
-querying of the catalogue can easily be incorporated into Python scripts.
+the catalogue. Querying of the catalogue can easily be incorporated into Python scripts.
 
 The module can also be used to create plots of pulsar period against period
 derivative ($P$ vs.\ $\dot{P}$ plots) using `matplotlib` [@matplotlib] as shown
