### 2018-02-07 #538 @mattpitkin
I've [added](https://github.com/mattpitkin/psrqpy/commit/983a57bae9cd6228da47f995c11748f6d38a76d8) a statement to the article and documentation about the intended users and purpose of the script.
