### 2018-02-07 3ca0483
Add explicit printed output example in the README and add a check in the travis build that the Crab pulsar's frequency rounds down to 29 Hz (this should be good for approximately 80 years)  - refs openjournals/joss-reviews#538
