### 2018-02-07 #538 @mattpitkin
I've [added](https://github.com/mattpitkin/psrqpy/commit/3ca04833859d9d7ffc98159e82762fcc1fae7e17) some actual output to the README file and added a test in the travis build that queries the frequency of the Crab pulsars and checks that it rounds down to 29 Hz (this should be good for another 80 years).
