### 2018-02-07 3246d0d
Update .travis.yml
fix missing '
