### 2018-02-07 3796f5e
Update .travis.yml
fix q->query
