### 2018-02-07 23cee57
Add an example to the documentation of returning a reference link  - refs openjournals/joss-reviews#538
