### 2018-02-07 d50086c
Add slightly expanded doc string to the QueryATNF and Pulsar classes. Also bump the version number with these additions.
 - refs openjournals/joss-reviews#538
