### 2018-02-07 #538 @mattpitkin
I've [added](https://github.com/mattpitkin/psrqpy/commit/23cee57d7b2f9e1cd1cadc70c0140e2c55e6ed80) an example to the documentation on getting an ADS reference link for a pulsar parameter. I've also [added](https://github.com/mattpitkin/psrqpy/commit/d50086c3a53910fd79d80f7cce9ca1457d31c4f7) a bit more (although to be honest not much more) documentation for the QueryATNF and Pulsar classes. After adding the additional docs I bumped the version number up to 0.4.6.
