### 2018-02-07 #538 @mattpitkin
@ygrange - I hope my above changes address you main issues. Let me know if there's anything else that you'd like to see added.
