### 2018-02-16 #538 @ygrange
Very minor comment: In the usage description the word "txet" appears.

I think for the rest, the paper is good to go as far as I am concerned.
