### 2018-02-16 25f4fdc
Update paper.md
 - fix typo
 - refs https://github.com/openjournals/joss-reviews/issues/538#issuecomment-366269675
