### 2018-02-16 #538 @mattpitkin
> Very minor comment: In the usage description the word "txet" appears.

Thanks, I've fixed to typo now.
