### 2018-02-18 #538 @ygrange
Yes I am! Sorry for the unclarity.

--
Y. Grange

> Op 18 feb. 2018 om 02:04 heeft Arfon Smith <notifications@github.com>
het volgende geschreven:
>
> @ygrange - just confirming we're good to accept here now?
>
