### 2018-02-18 #538 @mattpitkin
@arfon - sure, I've created an archive on Zenodo, with the DOI 10.5281/zenodo.1175303. [![DOI](https://zenodo.org/badge/110824647.svg)](https://zenodo.org/badge/latestdoi/110824647)
