### 2018-02-19 #538 @arfon
@ygrange - many thanks for your review here ✨

@mattpitkin - your paper is now accepted in JOSS and your DOI is https://doi.org/10.21105/joss.00538 ⚡️ 🚀 💥
